feat: Add a mutation event
A mutation event is now implemented in to the library. It uses the
MutationObserver. Event is useful when needing to check for any
mutations to the DOM or to a specific element via event delegation.

If needed to revert back pre-feature use commit id:
e4e6a5bff1b3728407b9095ba5fea0849a44c8db